+++ eras/allegra/impl/CHANGELOG.md
@@ -11,6 +11,10 @@
 * Add `EraTxLevel` instance
 * Remove deprecated `timelockScriptsTxAuxDataL`
 
+### `cddl`
+
+* Add `HasCDDL` type class instances for CDDL generation.
+
 ### `testlib`
 
 * Add CDDL definitions for int64 types: `int64`, `min_int64`, `max_int64`, `negative_int64`, `positive_int64`, `nonzero_int64`